pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asyncpg==0.30.0", # PostgreSQL async driver
    "bcrypt==4.0.1",
    "cffi==2.0.0",
    "click==8.3.1",
    "cryptography==46.0.3",
    "fastapi==0.121.2",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "pyjwt>=2.8,<3.0",
    "passlib==1.7.4",
    "pycparser==2.23",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydantic[email]==2.12.4",
    "python-dotenv==1.2.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "starlette==0.49.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.38.0",
    "alembic>=1.17.2",
    "python-multipart==0.0.20",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
]
